+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 
 ## Fixed
 
+- Surface CSV parse warnings in the UI ([#26](https://github.com/ButteredGroove/fac-flipper/issues/26))
 - Improve fetch errors with HTTP status details and clearer network failures ([#25](https://github.com/ButteredGroove/fac-flipper/issues/25))
 - Validate deck manifests for required fields ([#24](https://github.com/ButteredGroove/fac-flipper/issues/24))
 - Surface deck load failures in `selectDeck` ([#23](https://github.com/ButteredGroove/fac-flipper/issues/23))